requirements.txt
faiss_cpu==1.8.0.post1
fastapi==0.115.12
numpy==1.24.3
pandas==2.0.3
sentence_transformers==3.2.1
torch==2.4.1
